refactor(doctor): pay for the override check by extracting, not by bumping a baseline
check_source_size_budget refused the Phase-3 doctor check at +128 lines, correctly:
cmd_doctor.ts sits ~2,200 lines above the 1,500-line ceiling, so any net growth in it
is refused. Raising the baseline is what that gate's own message calls a defect rather
than a fix.

Paid out of the addition itself where possible — `_check_overrides` moved whole to
_cli/doctor_overrides.ts (-119) — and out of the cleanest pre-existing neighbour for
the residual: the install-mode detector and its marker constant moved to
_cli/doctor_install_mode.ts (-14). Both names are forwarded with `export ... from`, not
an import, because INSTALL_MODE_MARKER_RELATIVE is no longer referenced in cmd_doctor
and an unused import would be dropped — silently breaking whoever imports it from that
path.

Extracting a NEIGHBOUR rather than shaving the new feature is deliberate: the
alternative was compressing the new check's comments until the number fit, which buys
the same green from worse code.

Total excess 19,363 → 19,362. Net -1 past zero rather than exactly to it, because the
unit that paid is a unit and splitting it to hit a number would leave a worse file than
moving it. Baseline set to the EXACT live total — check_source_size_budget.test.ts
asserts baseline == tree total, so a baseline left one line loose is a red test.

Both new files are under the ceiling and contribute zero excess. The doctor's override
line is unchanged, and both touched test files stay green (15/15, 7/7).
